# docs/swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:3333","description":"Generated server url"}],"paths":{"/api/v1/atributo":{"post":{"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AtributoDto"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/Atributo"}}},"description":"OK"}},"tags":["atributo-controller"]}},"/api/v1/atributo/all":{"get":{"operationId":"getAllAtributos","responses":{"200":{"content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Atributo"}}}},"description":"OK"}},"tags":["atributo-controller"]}},"/api/v1/atributo/{id}":{"delete":{"operationId":"deleteAtributos","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"content":{"*/*":{"schema":{"type":"string"}}},"description":"OK"}},"tags":["atributo-controller"]},"get":{"operationId":"getByIdAtributos","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/Atributo"}}},"description":"OK"}},"tags":["atributo-controller"]},"put":{"operationId":"updateAtributos","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AtributoDto"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/Atributo"}}},"description":"OK"}},"tags":["atributo-controller"]}},"/api/v1/auth/login":{"post":{"operationId":"Login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/Token"}}},"description":"OK"}},"tags":["auth-controller"]}},"/api/v1/auth/register":{"post":{"operationId":"Register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}},"description":"OK"}},"tags":["auth-controller"]}},"/api/v1/auth/reset-password":{"post":{"operationId":"ResetPassword","requestBody":{"content":{"application/json":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"type":"string"}}},"description":"OK"}},"tags":["auth-controller"]}},"/api/v1/instituicao":{"post":{"operationId":"createInstituicaos","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InstituicaoDto"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/Instituicao"}}},"description":"OK"}},"tags":["instituicao-controller"]}},"/api/v1/instituicao/":{"get":{"operationId":"getAllInstituicaos","responses":{"200":{"content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Instituicao"}}}},"description":"OK"}},"tags":["instituicao-controller"]}},"/api/v1/instituicao/{id}":{"delete":{"operationId":"deleteInstituicaos","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}},"tags":["instituicao-controller"]},"get":{"operationId":"getByIdInstituicaos","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/Instituicao"}}},"description":"OK"}},"tags":["instituicao-controller"]},"put":{"operationId":"updateInstituicaos","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InstituicaoDto"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/Instituicao"}}},"description":"OK"}},"tags":["instituicao-controller"]}},"/api/v1/inventario-item":{"post":{"operationId":"createInventarioItem","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventarioItem"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/InventarioItem"}}},"description":"OK"}},"tags":["inventario-item-controller"]}},"/api/v1/inventario-item/all":{"get":{"operationId":"getAllInventarioItems","responses":{"200":{"content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InventarioItem"}}}},"description":"OK"}},"tags":["inventario-item-controller"]}},"/api/v1/inventario-item/user/{userId}":{"get":{"operationId":"getInventarioItemsByUser","parameters":[{"in":"path","name":"userId","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InventarioItem"}}}},"description":"OK"}},"tags":["inventario-item-controller"]}},"/api/v1/inventario-item/{id}":{"delete":{"operationId":"deleteInventarioItem","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"content":{"*/*":{"schema":{"type":"string"}}},"description":"OK"}},"tags":["inventario-item-controller"]},"get":{"operationId":"getByIdInventarioItem","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/InventarioItem"}}},"description":"OK"}},"tags":["inventario-item-controller"]},"put":{"operationId":"updateInventarioItem","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventarioItem"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/InventarioItem"}}},"description":"OK"}},"tags":["inventario-item-controller"]}},"/api/v1/item":{"post":{"operationId":"createItem","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemDto"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/Item"}}},"description":"OK"}},"tags":["item-controller"]}},"/api/v1/item/all":{"get":{"operationId":"getAllItems","responses":{"200":{"content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Item"}}}},"description":"OK"}},"tags":["item-controller"]}},"/api/v1/item/{id}":{"delete":{"operationId":"deleteItem","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"content":{"*/*":{"schema":{"type":"string"}}},"description":"OK"}},"tags":["item-controller"]},"get":{"operationId":"getByIdItem","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/Item"}}},"description":"OK"}},"tags":["item-controller"]},"put":{"operationId":"updateItem","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemDto"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/Item"}}},"description":"OK"}},"tags":["item-controller"]}},"/api/v1/loja-item":{"post":{"operationId":"createLojaItem","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LojaItem"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/LojaItem"}}},"description":"OK"}},"tags":["loja-item-controller"]}},"/api/v1/loja-item/all":{"get":{"operationId":"getAllLojaItems","responses":{"200":{"content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LojaItem"}}}},"description":"OK"}},"tags":["loja-item-controller"]}},"/api/v1/loja-item/{id}":{"delete":{"operationId":"deleteLojaItem","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"content":{"*/*":{"schema":{"type":"string"}}},"description":"OK"}},"tags":["loja-item-controller"]},"get":{"operationId":"getByIdLojaItem","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/LojaItem"}}},"description":"OK"}},"tags":["loja-item-controller"]},"put":{"operationId":"updateLojaItem","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LojaItem"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/LojaItem"}}},"description":"OK"}},"tags":["loja-item-controller"]}},"/api/v1/missions":{"get":{"operationId":"getAllMissoes","responses":{"200":{"content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Missoes"}}}},"description":"OK"}},"tags":["missoes-controller"]},"post":{"operationId":"createMissoes","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MissaoDto"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/Missoes"}}},"description":"OK"}},"tags":["missoes-controller"]}},"/api/v1/missions/{id}":{"delete":{"operationId":"deleteMissoes","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"content":{"*/*":{"schema":{"type":"string"}}},"description":"OK"}},"tags":["missoes-controller"]}},"/api/v1/missoes-submissoes":{"post":{"operationId":"createSubmissao","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MissoesSubmissoes"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/MissoesSubmissoes"}}},"description":"OK"}},"tags":["missoes-submissoes-controller"]}},"/api/v1/missoes-submissoes/all":{"get":{"operationId":"getAllSubmissoes","responses":{"200":{"content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MissoesSubmissoes"}}}},"description":"OK"}},"tags":["missoes-submissoes-controller"]}},"/api/v1/missoes-submissoes/user/{userId}":{"get":{"operationId":"getSubmissoesByUser","parameters":[{"in":"path","name":"userId","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MissoesSubmissoes"}}}},"description":"OK"}},"tags":["missoes-submissoes-controller"]}},"/api/v1/missoes-submissoes/{id}":{"delete":{"operationId":"deleteSubmissao","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"content":{"*/*":{"schema":{"type":"string"}}},"description":"OK"}},"tags":["missoes-submissoes-controller"]},"put":{"operationId":"updateSubmissao","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MissoesSubmissoes"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/MissoesSubmissoes"}}},"description":"OK"}},"tags":["missoes-submissoes-controller"]}},"/api/v1/nivel-definicao":{"post":{"operationId":"createNivelDefinicao","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NivelDefinicaoDto"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/NivelDefinicao"}}},"description":"OK"}},"tags":["nivel-definicao-controller"]}},"/api/v1/nivel-definicao/all":{"get":{"operationId":"getAllNivelDefinicao","responses":{"200":{"content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NivelDefinicao"}}}},"description":"OK"}},"tags":["nivel-definicao-controller"]}},"/api/v1/nivel-definicao/{id}":{"delete":{"operationId":"deleteNivelDefinicao","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"content":{"*/*":{"schema":{"type":"string"}}},"description":"OK"}},"tags":["nivel-definicao-controller"]},"get":{"operationId":"getByIdNivelDefinicao","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/NivelDefinicao"}}},"description":"OK"}},"tags":["nivel-definicao-controller"]},"put":{"operationId":"updateNivelDefinicao","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NivelDefinicaoDto"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/NivelDefinicao"}}},"description":"OK"}},"tags":["nivel-definicao-controller"]}},"/api/v1/questionario":{"post":{"operationId":"createQuestionario","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Questionario"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/Questionario"}}},"description":"OK"}},"tags":["questionario-controller"]}},"/api/v1/questionario/all":{"get":{"operationId":"getAllQuestionarios","responses":{"200":{"content":{"*/*":{"schema":{"type":"object"}}},"description":"OK"}},"tags":["questionario-controller"]}},"/api/v1/questionario/missao/{missaoId}":{"get":{"operationId":"getQuestionarioByMissao","parameters":[{"in":"path","name":"missaoId","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/Questionario"}}},"description":"OK"}},"tags":["questionario-controller"]}},"/api/v1/questionario/{id}":{"delete":{"operationId":"deleteQuestionario","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"content":{"*/*":{"schema":{"type":"string"}}},"description":"OK"}},"tags":["questionario-controller"]},"get":{"operationId":"getByIdQuestionario","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/Questionario"}}},"description":"OK"}},"tags":["questionario-controller"]},"put":{"operationId":"updateQuestionario","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Questionario"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/Questionario"}}},"description":"OK"}},"tags":["questionario-controller"]}},"/api/v1/turma":{"post":{"operationId":"createTurma","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TurmaDto"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/Turma"}}},"description":"OK"}},"tags":["turma-controller"]}},"/api/v1/turma/all":{"get":{"operationId":"getAllTurma","responses":{"200":{"content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Turma"}}}},"description":"OK"}},"tags":["turma-controller"]}},"/api/v1/turma/qt":{"get":{"operationId":"getTurmas","responses":{"200":{"content":{"*/*":{"schema":{"type":"integer","format":"int64"}}},"description":"OK"}},"tags":["turma-controller"]}},"/api/v1/turma/{id}":{"delete":{"operationId":"deleteTurma","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"content":{"*/*":{"schema":{"type":"string"}}},"description":"OK"}},"tags":["turma-controller"]},"put":{"operationId":"updateTurma","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TurmaDto"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/Turma"}}},"description":"OK"}},"tags":["turma-controller"]}},"/api/v1/user":{"get":{"operationId":"getUserById","parameters":[{"in":"query","name":"id","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}},"description":"OK"}},"tags":["user-controller"]},"post":{"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}},"description":"OK"}},"tags":["user-controller"]}},"/api/v1/user/all":{"get":{"operationId":"getUser","responses":{"200":{"content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}},"description":"OK"}},"tags":["user-controller"]}},"/api/v1/user/qt":{"get":{"operationId":"getQtUsers","responses":{"200":{"content":{"*/*":{"schema":{"type":"integer","format":"int64"}}},"description":"OK"}},"tags":["user-controller"]}}},"components":{"schemas":{"Atributo":{"type":"object","properties":{"diamantes":{"type":"integer","format":"int32"},"esmeraldas":{"type":"integer","format":"int32"},"id":{"type":"integer","format":"int64"},"moedas":{"type":"integer","format":"int32"},"rubys":{"type":"integer","format":"int32"},"xp":{"type":"integer","format":"int32"}}},"AtributoDto":{"type":"object","properties":{"diamantes":{"type":"integer","format":"int32"},"esmeraldas":{"type":"integer","format":"int32"},"moedas":{"type":"integer","format":"int32"},"rubys":{"type":"integer","format":"int32"},"xp":{"type":"integer","format":"int32"}}},"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"Instituicao":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nome":{"type":"string"},"turma":{"type":"array","items":{"$ref":"#/components/schemas/Turma"}}}},"InstituicaoDto":{"type":"object","properties":{"nome":{"type":"string"}}},"Inventario":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"user":{"$ref":"#/components/schemas/User"}}},"InventarioItem":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"inventario":{"$ref":"#/components/schemas/Inventario"},"item":{"$ref":"#/components/schemas/Item"},"quantidade":{"type":"integer","format":"int32"}}},"Item":{"type":"object","properties":{"descricao":{"type":"string"},"id":{"type":"integer","format":"int32"},"nome":{"type":"string"}}},"ItemDto":{"type":"object","properties":{"descricao":{"type":"string"},"nome":{"type":"string"}}},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"senha":{"type":"string"}}},"LojaItem":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"item":{"$ref":"#/components/schemas/Item"},"preco":{"type":"integer","format":"int32"},"tipoMoeda":{"type":"string"}}},"MissaoDto":{"type":"object","properties":{"descricao":{"type":"string"},"diamantesRecompensa":{"type":"integer","format":"int32"},"esmeraldasRecompensa":{"type":"integer","format":"int32"},"moedasRecompensa":{"type":"integer","format":"int32"},"nome":{"type":"string"},"rubysRecompensa":{"type":"integer","format":"int32"},"turmaId":{"type":"integer","format":"int64"},"xpRecompensa":{"type":"integer","format":"int32"}}},"Missoes":{"type":"object","properties":{"conteudo":{"type":"string"},"descricao":{"type":"string"},"diamantesRecompensa":{"type":"integer","format":"int32"},"esmeraldasRecompensa":{"type":"integer","format":"int32"},"id":{"type":"integer","format":"int32"},"moedasRecompensa":{"type":"integer","format":"int32"},"nome":{"type":"string"},"rubysRecompensa":{"type":"integer","format":"int32"},"xpRecompensa":{"type":"integer","format":"int32"}}},"MissoesSubmissoes":{"type":"object","properties":{"arquivoUrl":{"type":"string"},"dataAvaliacao":{"type":"string","format":"date-time"},"dataEnvio":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int32"},"missao":{"$ref":"#/components/schemas/Missoes"},"observacao":{"type":"string"},"status":{"type":"string"},"user":{"$ref":"#/components/schemas/User"}}},"NivelDefinicao":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nome":{"type":"string"},"pontosNecessarios":{"type":"integer","format":"int32"}}},"NivelDefinicaoDto":{"type":"object","properties":{"nome":{"type":"string"},"pontosNecessarios":{"type":"integer","format":"int32"}}},"Questionario":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"missao":{"$ref":"#/components/schemas/Missoes"},"questoes":{"type":"array","items":{"$ref":"#/components/schemas/Questoes"}},"titulo":{"type":"string"}}},"Questoes":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"opcaoA":{"type":"string"},"opcaoB":{"type":"string"},"opcaoC":{"type":"string"},"opcaoD":{"type":"string"},"pergunta":{"type":"string"},"respostaCorreta":{"type":"string"}}},"Token":{"type":"object","properties":{"github":{"type":"string"},"nome":{"type":"string"},"token":{"type":"string"}}},"Turma":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"periodo":{"type":"string"}}},"TurmaDto":{"type":"object","properties":{"instituicao":{"type":"integer","format":"int32"},"periodo":{"type":"string"},"turma":{"type":"string"}}},"User":{"type":"object","properties":{"accountNonExpired":{"type":"boolean"},"accountNonLocked":{"type":"boolean"},"atributo":{"$ref":"#/components/schemas/Atributo"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"credentialsNonExpired":{"type":"boolean"},"email":{"type":"string"},"enabled":{"type":"boolean"},"github":{"type":"string"},"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["ADMIN","USER"]},"senha":{"type":"string"},"turma":{"$ref":"#/components/schemas/Turma"},"username":{"type":"string"}}},"UserDto":{"type":"object","properties":{"email":{"type":"string"},"github":{"type":"string"},"nome":{"type":"string"},"senha":{"type":"string"},"turma":{"type":"integer","format":"int64"}}}}}}